[mod] アカウント確認画面でaxiosによるログイン処理を実行する形に修正   - アカウント登録画面では送信確認ボタン押下時に、入力データをvuexに渡して画面遷移を行う
