+++ rails_app/app/javascript/components/store/StoreAccountConfirm.vue
@@ -12,6 +12,9 @@
 
       <div class="row g-5 flex justify-center">
         <div class="col-md-7 col-lg-8">
+          <div class="col-12 alert alert-danger" v-show="hasError">
+            <p>{{errorMessage}}</p>
+          </div>
           <form class="needs-validation" @submit="validate" novalidate>
             <div class="row g-3">
               <div class="col-12">
@@ -97,13 +100,17 @@
 </template>
 
 <script>
+import Router from "../../router/router";
 import StoreHeader from "../layout/StoreHeader.vue"
 import { mapGetters } from 'vuex'
+import axios from "axios";
 
 export default {
   data: function () {
     return {
-      storeData: []
+      storeData: [],
+      hasError: false,
+      errorMessage: "",
     }
   },
 
@@ -112,6 +119,36 @@ export default {
   },
 
   methods: {
+        // 店舗情報を送信する
+        async submit() {
+            this.loading = true;
+            await axios
+                .post("/api/v1/store_auth/", this.storeData)
+                .then(response => {
+
+                    // Vuex store
+                    this.$store.dispatch('storeAuth/updateLogin', true)
+
+                    Router.push("/store_account_confirm");
+                })
+                .catch(e => {
+                    if (e.response) {
+                        this.hasError = true;
+                        this.errorMessage = e.response.data.errors.full_messages[0];
+
+                        this.showErrorMessage(e)
+
+                        console.log(e.response.data);
+                        console.log(e.response.status);
+                        console.log(e.response.headers);
+                    } else {
+                        console.log("Error", e.message);
+                    }
+                })
+                .finally(() => {
+                    this.loading = false;
+                });
+        },
     validate(event) {
       if (!event.target.checkValidity()) {
         event.preventDefault()
